api/pyproject.toml
[tool.black]
line-length = 79

[tool.pydocstyle]
convention = "google"
match = ".*\\.py"
match-dir = "(?!migrations).*"
add-ignore = "D107,D104,D106"

[tool.poetry]
name = "nutrition"
version = "0.1.0"
description = ""
authors = ["Luis Muñoz <luis.munoz.villarreal@gmail.com>"]
packages = [
    { include = "apps" },
    { include = "config" },
    { include = "manage.py" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
django-admin-sortable2 = "2.1.5"
django-compressor = "4.3.1"
django-cors-headers = "3.14.0"
django-extensions = "3.2.1"
django-nested-inline = "0.4.6"
django = "4.2"
pint = "0.20.1"

[tool.poetry.group.dev.dependencies]
black = "23.3.0"
flake8 = "6.0.0"
flake8-isort = "6.0.0"
ipython = "8.12.0"
ipdb = "0.13.13"
mypy = "1.1.1"
pydocstyle = "6.3.0"
pylint-django = "2.5.3"
pylint = "2.17.1"
pytest-cov = "4.0.0"
pytest-django = "4.5.2"
pytest-factoryboy = "2.5.1"
pytest-mock = "3.10.0"
pytest-xdist = "3.2.1"
pytest = "7.2.2"
tox = "4.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
